refactor: 열차 운임 계산 클래스 네이밍, 패키지 위치 변경 (#114)
