+++ src/test/java/com/sudo/raillo/support/helper/OrderTestHelper.java
@@ -1,6 +1,6 @@
 package com.sudo.raillo.support.helper;
 
-import com.sudo.raillo.booking.application.service.FareCalculationService;
+import com.sudo.raillo.train.application.calculator.FareCalculator;
 import com.sudo.raillo.booking.domain.type.PassengerType;
 import com.sudo.raillo.member.domain.Member;
 import com.sudo.raillo.order.domain.Order;
@@ -16,8 +16,6 @@ import com.sudo.raillo.train.domain.type.CarType;
 import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Lazy;
@@ -30,10 +28,10 @@ import org.springframework.transaction.annotation.Transactional;
 public class OrderTestHelper {
 
 	private final TrainTestHelper trainTestHelper;
-	private final FareCalculationService fareCalculationService;
 	private final OrderRepository orderRepository;
 	private final OrderBookingRepository orderBookingRepository;
 	private final OrderSeatBookingRepository orderSeatBookingRepository;
+	private final FareCalculator fareCalculator;
 
 	@Lazy
 	@Autowired
@@ -128,7 +126,7 @@ public class OrderTestHelper {
 
 		List<OrderSeatBooking> toSave = builder.seatWithPassengerTypes.stream()
 			.map(sp -> {
-				BigDecimal fare = fareCalculationService.calculateFare(
+				BigDecimal fare = fareCalculator.calculateFare(
 					departureStationId,
 					arrivalStationId,
 					sp.passengerType(),
@@ -290,7 +288,7 @@ public class OrderTestHelper {
 
 			// 각 CarType별로 운임 계산 후 합산
 			return seatWithPassengerTypes.stream()
-				.map(sp -> fareCalculationService.calculateFare(
+				.map(sp -> fareCalculator.calculateFare(
 					departureScheduleStop.getStation().getId(),
 					arrivalScheduleStop.getStation().getId(),
 					sp.passengerType,
